requirements.txt
flask-restplus==0.12.1
Flask-PyMongo==2.2.0
Flask==1.0.2
Flask-Mail==0.9.1
passlib==1.7.1
